Project structure updated and successful solved the cart-pole game
The file train.py is a very first tentative to implement the first visit Monte Carlo algorithm
for solving the problem of the cart pole offered from: https://gym.openai.com/.

From https://github.com/openai/gym/blob/master/gym/envs/classic_control/cartpole.py
the problem is considered solved when the average return is greater than or equal to
195.0 over 100 consecutive trials.

In order to keep up exploration during the training a random action is selected.
The algorithm starts to become greedy when the episodes are close to the pre-selected EPISODES.

A score of 196.6 was obtained after 1200 episodes.